Rename boot module to application
Renames the entrypoint module from `boot` to `application`, updates the main package to `pt.estga`, and adjusts the Maven module list and Spring Boot main class accordingly. Also updates the README run instructions to match the new application class name.
