+++ application/src/main/java/pt/estga/RedirectController.java
@@ -1,0 +1,15 @@
+package pt.estga;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping()
+public class RedirectController {
+
+    @GetMapping("/")
+    public String redirectToSwagger() {
+        return "redirect:/swagger-ui.html";
+    }
+}
