+++ application/src/main/java/pt/estga/config/TimeConfiguration.java
@@ -1,0 +1,14 @@
+package pt.estga.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import java.time.Clock;
+
+@Configuration
+public class TimeConfiguration {
+
+    @Bean
+    public Clock systemClock() {
+        return Clock.systemUTC();
+    }
+}
