+++ application/src/main/java/pt/estga/dtos/AuthResponse.java
@@ -1,0 +1,8 @@
+package pt.estga.dtos;
+
+public record AuthResponse(
+        String accessToken,
+        String refreshToken,
+        String role,
+        Long userId
+) {}
