+++ application/src/main/java/pt/estga/dtos/GoogleAuthRequest.java
@@ -1,0 +1,7 @@
+package pt.estga.dtos;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record GoogleAuthRequest(
+        @NotBlank String token
+) {}
