+++ application/src/main/java/pt/estga/dtos/LoginRequest.java
@@ -1,0 +1,8 @@
+package pt.estga.dtos;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record LoginRequest(
+        @NotBlank String username,
+        @NotBlank String password
+) {}
