+++ application/src/main/java/pt/estga/dtos/RefreshTokenRequest.java
@@ -1,0 +1,7 @@
+package pt.estga.dtos;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record RefreshTokenRequest(
+        @NotBlank String refreshToken
+) {}
