+++ application/src/main/resources/db/migration/V1__Create_extensions.sql
