--- boot/src/main/java/pt/estga/boot/AuthController.java
@@ -1,49 +1,0 @@
-package pt.estga.boot;
-
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import pt.estga.boot.dtos.AuthResponse;
-import pt.estga.boot.dtos.GoogleAuthRequest;
-import pt.estga.boot.dtos.LoginRequest;
-import pt.estga.boot.dtos.RefreshTokenRequest;
-import pt.estga.boot.services.AuthService;
-
-@RestController
-@RequestMapping("/api/v1/auth")
-@RequiredArgsConstructor
-@Tag(name = "Authentication", description = "Authentication endpoints")
-public class AuthController {
-
-    private final AuthService authService;
-
-    @PostMapping("/authenticate")
-    @Operation(summary = "Authenticate with username and password")
-    public ResponseEntity<AuthResponse> authenticate(@Valid @RequestBody LoginRequest request) {
-        return ResponseEntity.ok(authService.authenticate(request.username(), request.password()));
-    }
-
-    @PostMapping("/google")
-    @Operation(summary = "Authenticate with Google ID token")
-    public ResponseEntity<AuthResponse> google(@Valid @RequestBody GoogleAuthRequest request) {
-        return ResponseEntity.ok(authService.authenticateWithGoogle(request.token()));
-    }
-
-    @PostMapping("/refresh")
-    @Operation(summary = "Refresh access token")
-    public ResponseEntity<AuthResponse> refresh(@Valid @RequestBody RefreshTokenRequest request) {
-        return ResponseEntity.ok(authService.refresh(request.refreshToken()));
-    }
-
-    @PostMapping("/logout")
-    @Operation(summary = "Logout")
-    public ResponseEntity<Void> logout() {
-        return ResponseEntity.ok().build();
-    }
-}
