--- boot/src/main/java/pt/estga/boot/RedirectController.java
@@ -1,15 +1,0 @@
-package pt.estga.boot;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping()
-public class RedirectController {
-
-    @GetMapping("/")
-    public String redirectToSwagger() {
-        return "redirect:/swagger-ui.html";
-    }
-}
