--- boot/src/main/java/pt/estga/boot/config/AuditingConfig.java
@@ -1,18 +1,0 @@
-package pt.estga.boot.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.domain.AuditorAware;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import pt.estga.commoncore.utils.SecurityUtils;
-
-@Configuration
-@EnableJpaAuditing(auditorAwareRef = "auditorAware")
-public class AuditingConfig {
-
-    @Bean
-    AuditorAware<Long> auditorAware() {
-        return () -> SecurityUtils.currentPrincipal()
-                .map(p -> p.getId());
-    }
-}
