--- boot/src/main/java/pt/estga/boot/config/TimeConfiguration.java
@@ -1,14 +1,0 @@
-package pt.estga.boot.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import java.time.Clock;
-
-@Configuration
-public class TimeConfiguration {
-
-    @Bean
-    public Clock systemClock() {
-        return Clock.systemUTC();
-    }
-}
