--- boot/src/main/java/pt/estga/boot/dtos/AuthResponse.java
@@ -1,8 +1,0 @@
-package pt.estga.boot.dtos;
-
-public record AuthResponse(
-        String accessToken,
-        String refreshToken,
-        String role,
-        Long userId
-) {}
