--- boot/src/main/java/pt/estga/boot/dtos/GoogleAuthRequest.java
@@ -1,7 +1,0 @@
-package pt.estga.boot.dtos;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record GoogleAuthRequest(
-        @NotBlank String token
-) {}
