--- boot/src/main/java/pt/estga/boot/dtos/LoginRequest.java
@@ -1,8 +1,0 @@
-package pt.estga.boot.dtos;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record LoginRequest(
-        @NotBlank String username,
-        @NotBlank String password
-) {}
