--- boot/src/main/java/pt/estga/boot/dtos/RefreshTokenRequest.java
@@ -1,7 +1,0 @@
-package pt.estga.boot.dtos;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record RefreshTokenRequest(
-        @NotBlank String refreshToken
-) {}
